Move AlertManager into every screen in my MainTabs
So that it will be within a Navigator and will appear in all our main screens.
