+++ src/components/AlertManager/AlertManager.jsx
@@ -1,91 +1,17 @@
 // Third-party dependencies
 import React from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import { useSelector } from 'react-redux'
 import Portal from '@burstware/react-native-portal'
-import { isEmpty, map } from 'lodash'
-import OneSignal from 'react-native-onesignal'
-import { Alert } from 'react-native'
-import { ONESIGNAL_APP_ID } from '@env'
+import { isEmpty } from 'lodash'
 
 // In-house dependencies
 import { ModalView, ModalContainer, AlertModal } from '../Modals'
 import { ALERT_TYPE, ALERT_STATUS } from '../../constants'
 import colors from '../../resources/colors'
 import { getAlerts } from '../../redux/selectors'
-import { setAlerts } from '../../redux/slices/alertsSlice'
-import Logger from '../../services/Logger'
-
-const logger = new Logger('AlertManager')
 
 function AlertManager() {
-  const dispatch = useDispatch()
-
   const alerts = useSelector(getAlerts)
-
-  const incidentTypes = ['Overdose', 'Unsafe Guest', 'Accidental', 'Mental Health', 'Other']
-
-  // OneSignal Init Code
-  OneSignal.setLogLevel(6, 0)
-  OneSignal.setAppId(ONESIGNAL_APP_ID)
-
-  // Prompt for push on iOS
-  OneSignal.promptForPushNotificationsWithUserResponse(response => {
-    logger.debug('OneSignal: prompt response:', response)
-  })
-
-  function displayAlerts() {
-    // TODO get from API call
-    dispatch(
-      setAlerts([
-        {
-          id: 'guid-123-abc',
-          alertStatus: ALERT_STATUS.NEW,
-          location: 'Room 302',
-          alertType: ALERT_TYPE.BUTTONS_NOT_URGENT,
-          categories: ['Safer Use', 'Overdose', 'Mental Health', 'Other'],
-        },
-        {
-          id: 'guid-123-abc2',
-          alertStatus: ALERT_STATUS.REMINDING,
-          location: 'Room 302',
-          alertType: ALERT_TYPE.SENSOR_DURATION,
-          categories: ['Safer Use', 'Overdose', 'Mental Health', 'Other'],
-        },
-        {
-          id: 'guid-123-abc3',
-          alertStatus: ALERT_STATUS.RESPONDING,
-          location: 'Room 302',
-          alertType: ALERT_TYPE.BUTTONS_NOT_URGENT,
-          categories: ['Safer Use', 'Overdose', 'Mental Health', 'Other'],
-        },
-        {
-          id: 'guid-123-abc4',
-          alertStatus: ALERT_STATUS.REPORTING,
-          location: 'Room 302',
-          alertType: ALERT_TYPE.BUTTONS_NOT_URGENT,
-          categories: ['Safer Use', 'Overdose', 'Mental Health', 'Other'],
-        },
-      ]),
-    )
-  }
-
-  // Method for handling notifications opened
-  OneSignal.setNotificationOpenedHandler(notification => {
-    logger.debug('OneSignal: notification opened:', notification)
-
-    displayAlerts()
-  })
-
-  // Method for handling notifications received while app in foreground
-  OneSignal.setNotificationWillShowInForegroundHandler(notificationReceivedEvent => {
-    const notification = notificationReceivedEvent.getNotification()
-    logger.info('OneSignal: notification: ', notification)
-
-    // Don't show the native notification
-    notificationReceivedEvent.complete(null)
-
-    displayAlerts()
-  })
 
   function renderAlert(alert) {
     const borderColor =
@@ -107,6 +33,7 @@ function AlertManager() {
       </ModalView>
     )
   }
+
   return (
     <Portal>
       <ModalContainer isModalVisible={!isEmpty(alerts)}>{alerts.map(renderAlert)}</ModalContainer>
