+++ src/screens/AlertHistoryScreen.jsx
@@ -12,6 +12,7 @@ import { useSafeHandler } from '../hooks'
 import SCREEN from '../navigation/ScreensEnum'
 import { formatTimeString, isSameYearMonthDate } from '../helpers'
 import colors from '../resources/colors'
+import AlertManager from '../components/AlertManager'
 import Logger from '../services/Logger'
 
 const logger = new Logger('AlertHistoryScreen')
@@ -119,6 +120,7 @@ function AlertHistoryScreen() {
 
   return (
     <>
+      <AlertManager />
       <StatusBar barStyle="dark-content" />
       <SafeAreaView style={styles.container}>
         {fireGetHistoricAlertsRequestOptions.isFiring && (
