+++ src/screens/HomeScreen.jsx
@@ -14,6 +14,7 @@ import AlertApiService from '../services/AlertApiService'
 import SCREEN from '../navigation/ScreensEnum'
 import HomeScreenInstructions from '../components/HomeScreenInstructions'
 import HomeScreenNotificationsButton from '../components/HomeScreenNotificationsButton'
+import AlertManager from '../components/AlertManager'
 
 const styles = StyleSheet.create({
   scrollView: {
@@ -78,6 +79,7 @@ function HomeScreen() {
 
   return (
     <>
+      <AlertManager />
       <StatusBar barStyle="dark-content" />
       <SafeAreaView>
         <ScrollView contentInsetAdjustmentBehavior="automatic" style={styles.scrollView}>
