+++ src/screens/SplashScreen.jsx
@@ -3,14 +3,17 @@ import React, { useEffect } from 'react'
 import { ActivityIndicator, SafeAreaView, StyleSheet, StatusBar, Text } from 'react-native'
 import { useNavigation } from '@react-navigation/native'
 import { useDispatch } from 'react-redux'
-import { BUTTONS_BASE_URL, SENSOR_BASE_URL } from '@env'
+import OneSignal from 'react-native-onesignal'
+import { BUTTONS_BASE_URL, SENSOR_BASE_URL, ONESIGNAL_APP_ID } from '@env'
 
 // In-house dependencies
+import { ALERT_TYPE, ALERT_STATUS } from '../constants'
 import { useSafeHandler } from '../hooks'
 import { setIsButtonsLocation, setIsSensorLocation, setLocationName } from '../redux/slices/locationSlice'
 import AlertApiService from '../services/AlertApiService'
 import colors from '../resources/colors'
 import SCREEN from '../navigation/ScreensEnum'
+import { setAlerts } from '../redux/slices/alertsSlice'
 import Logger from '../services/Logger'
 
 const logger = new Logger('SplashScreen')
@@ -74,6 +77,69 @@ function SplashScreen() {
       rollbackScreen: SCREEN.ONBOARDING,
     })
   }, [])
+
+  // OneSignal Init Code
+  OneSignal.setLogLevel(6, 0)
+  OneSignal.setAppId(ONESIGNAL_APP_ID)
+
+  // Prompt for push on iOS
+  OneSignal.promptForPushNotificationsWithUserResponse(response => {
+    logger.debug('OneSignal: prompt response:', response)
+  })
+
+  function displayAlerts() {
+    // TODO get from API call
+    dispatch(
+      setAlerts([
+        {
+          id: 'guid-123-abc',
+          alertStatus: ALERT_STATUS.NEW,
+          location: 'Room 302',
+          alertType: ALERT_TYPE.BUTTONS_NOT_URGENT,
+          categories: ['Safer Use', 'Overdose', 'Mental Health', 'Other'],
+        },
+        {
+          id: 'guid-123-abc2',
+          alertStatus: ALERT_STATUS.REMINDING,
+          location: 'Room 302',
+          alertType: ALERT_TYPE.SENSOR_DURATION,
+          categories: ['Safer Use', 'Overdose', 'Mental Health', 'Other'],
+        },
+        {
+          id: 'guid-123-abc3',
+          alertStatus: ALERT_STATUS.RESPONDING,
+          location: 'Room 302',
+          alertType: ALERT_TYPE.BUTTONS_NOT_URGENT,
+          categories: ['Safer Use', 'Overdose', 'Mental Health', 'Other'],
+        },
+        {
+          id: 'guid-123-abc4',
+          alertStatus: ALERT_STATUS.REPORTING,
+          location: 'Room 302',
+          alertType: ALERT_TYPE.BUTTONS_NOT_URGENT,
+          categories: ['Safer Use', 'Overdose', 'Mental Health', 'Other'],
+        },
+      ]),
+    )
+  }
+
+  // Method for handling notifications opened
+  OneSignal.setNotificationOpenedHandler(notification => {
+    logger.debug('OneSignal: notification opened:', notification)
+
+    displayAlerts()
+  })
+
+  // Method for handling notifications received while app in foreground
+  OneSignal.setNotificationWillShowInForegroundHandler(notificationReceivedEvent => {
+    const notification = notificationReceivedEvent.getNotification()
+    logger.info('OneSignal: notification: ', notification)
+
+    // Don't show the native notification
+    notificationReceivedEvent.complete(null)
+
+    displayAlerts()
+  })
 
   return (
     <>
